Setup for parallel tests
