+++ test/global.test.ts
@@ -6,7 +6,7 @@
 import { TestOutputChannel, TestUserInput } from '@microsoft/vscode-azext-dev';
 import * as assert from 'assert';
 import * as vscode from 'vscode';
-import { ext, registerOnActionStartHandler } from '../extension.bundle';
+import { ext, registerAzureUtilsExtensionVariables, registerOnActionStartHandler, registerUIExtensionVariables } from '../extension.bundle';
 
 const longRunningLocalTestsEnabled: boolean = !/^(false|0)?$/i.test(process.env.AzCode_EnableLongRunningTestsLocal || '');
 const longRunningRemoteTestsEnabled: boolean = !/^(false|0)?$/i.test(process.env.AzCode_UseAzureFederatedCredentials || '');
@@ -25,6 +25,8 @@ suiteSetup(async function (this: Mocha.Context): Promise<void> {
     }
 
     ext.outputChannel = new TestOutputChannel();
+    registerUIExtensionVariables(ext);
+    registerAzureUtilsExtensionVariables(ext);
 
     registerOnActionStartHandler(context => {
         // Use `TestUserInput` by default so we get an error if an unexpected call to `context.ui` occurs, rather than timing out
